Port oauth application model tests
