+++ tests/ignore/test_models.py
@@ -55,103 +55,6 @@ from website.views import find_bookmark_collection
 GUID_FACTORIES = UserFactory, NodeFactory, ProjectFactory
 
 
-class TestGUID(OsfTestCase):
-
-    def setUp(self):
-        super(TestGUID, self).setUp()
-        self.records = {}
-        for factory in GUID_FACTORIES:
-            record = factory()
-            self.records[record._name] = record
-
-    def test_guid(self):
-
-        for record in self.records.values():
-
-            record_guid = Guid.load(record._primary_key)
-
-            # GUID must exist
-            assert_false(record_guid is None)
-
-            # Primary keys of GUID and record must be the same
-            assert_equal(
-                record_guid._primary_key,
-                record._primary_key
-            )
-
-            # GUID must refer to record
-            assert_equal(
-                record_guid.referent,
-                record
-            )
-
-
-class TestApiOAuth2Application(OsfTestCase):
-    def setUp(self):
-        super(TestApiOAuth2Application, self).setUp()
-        self.api_app = ApiOAuth2ApplicationFactory()
-
-    def test_must_have_owner(self):
-        with assert_raises(ValidationError):
-            api_app = ApiOAuth2ApplicationFactory(owner=None)
-            api_app.save()
-
-    def test_client_id_auto_populates(self):
-        assert_greater(len(self.api_app.client_id), 0)
-
-    def test_client_secret_auto_populates(self):
-        assert_greater(len(self.api_app.client_secret), 0)
-
-    def test_new_app_is_not_flagged_as_deleted(self):
-        assert_true(self.api_app.is_active)
-
-    def test_cant_edit_creation_date(self):
-        with assert_raises(AttributeError):
-            self.api_app.date_created = timezone.now()
-
-    def test_invalid_home_url_raises_exception(self):
-        with assert_raises(ValidationError):
-            api_app = ApiOAuth2ApplicationFactory(home_url="Totally not a URL")
-            api_app.save()
-
-    def test_invalid_callback_url_raises_exception(self):
-        with assert_raises(ValidationError):
-            api_app = ApiOAuth2ApplicationFactory(callback_url="itms://itunes.apple.com/us/app/apple-store/id375380948?mt=8")
-            api_app.save()
-
-    def test_name_cannot_be_blank(self):
-        with assert_raises(ValidationError):
-            api_app = ApiOAuth2ApplicationFactory(name='')
-            api_app.save()
-
-    def test_long_name_raises_exception(self):
-        long_name = ('JohnJacobJingelheimerSchmidtHisNameIsMyN' * 5) + 'a'
-        with assert_raises(ValidationError):
-            api_app = ApiOAuth2ApplicationFactory(name=long_name)
-            api_app.save()
-
-    def test_long_description_raises_exception(self):
-        long_desc = ('JohnJacobJingelheimerSchmidtHisNameIsMyN' * 25) + 'a'
-        with assert_raises(ValidationError):
-            api_app = ApiOAuth2ApplicationFactory(description=long_desc)
-            api_app.save()
-
-    @mock.patch('framework.auth.cas.CasClient.revoke_application_tokens')
-    def test_active_set_to_false_upon_successful_deletion(self, mock_method):
-        mock_method.return_value(True)
-        self.api_app.deactivate(save=True)
-        self.api_app.reload()
-        assert_false(self.api_app.is_active)
-
-    @mock.patch('framework.auth.cas.CasClient.revoke_application_tokens')
-    def test_active_remains_true_when_cas_token_deletion_fails(self, mock_method):
-        mock_method.side_effect = cas.CasHTTPError("CAS can't revoke tokens", 400, 'blank', 'blank')
-        with assert_raises(cas.CasHTTPError):
-            self.api_app.deactivate(save=True)
-        self.api_app.reload()
-        assert_true(self.api_app.is_active)
-
-
 class TestNodeWikiPage(OsfTestCase):
 
     def setUp(self):
